requirements.txt
# Core Flask Framework
Flask==2.3.3
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0

# AI and Machine Learning
transformers>=4.50.0
torch>=2.0.0
accelerate>=0.20.0
numpy==2.2.6
opencv-python==4.12.0.88
Pillow==11.2.1

# Database and Caching
redis==6.2.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP and API
requests==2.31.0
python-multipart==0.0.20
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10

# Additional Dependencies
typing_extensions==4.13.2
typing-inspection==0.4.1
pydantic==2.11.5
pydantic_core==2.33.2
tenacity==8.5.0
sniffio==1.3.1
anyio==4.9.0
httpcore==1.0.9
httpx==0.28.1
websockets==15.0.1 